+++ src/Products/MeetingCommunes/tests/testCustomMeetingItem.py
@@ -3,6 +3,7 @@
 from DateTime import DateTime
 from Products.MeetingCommunes.config import FINANCE_ADVICES_COLLECTION_ID
 from Products.MeetingCommunes.tests.MeetingCommunesTestCase import MeetingCommunesTestCase
+from Products.PloneMeeting.indexes import REAL_ORG_UID_PATTERN
 
 
 class testCustomMeetingItem(MeetingCommunesTestCase):
@@ -94,7 +95,7 @@ class testCustomMeetingItem(MeetingCommunesTestCase):
              'o': 'plone.app.querystring.operation.selection.is',
              'v': ['delay_row_id__unique_id_001_2',
                    'delay_row_id__unique_id_002_2',
-                   'real_org_uid__{0}'.format(self.developers_uid),]}
+                   REAL_ORG_UID_PATTERN.format(self.developers_uid)]}
         ], )
         # create an item without finance advice
         self.changeUser('pmManager')
@@ -163,6 +164,27 @@ class testCustomMeetingItem(MeetingCommunesTestCase):
         form.handleSaveRemoveAdviceInheritance(form, None)
         self.assertTrue(clonedItem2.adapted().showFinanceAdviceTemplate())
         self.assertEqual(clonedItem2.adapted().getFinanceAdviceId(), self.developers_uid)
+
+        # auto asked advice
+        custom_advisers = list(cfg.getCustomAdvisers()) + [{
+            'row_id': 'unique_id_004',
+            'org': self.vendors_uid,
+            'for_item_created_from': today,
+            'gives_auto_advice_on': 'python: True',
+            'gives_auto_advice_on_help_message': '',
+            'is_linked_to_previous_row': '0'}]
+        cfg.setCustomAdvisers(custom_advisers)
+        auto_item = self.create('MeetingItem')
+        self.assertTrue(self.vendors_uid in auto_item.adviceIndex)
+        # not a finance advice for now
+        self.assertFalse(auto_item.adapted().getFinanceAdviceId())
+        self.assertFalse(auto_item.adapted().showFinanceAdviceTemplate())
+        # make is recognized as finance advice
+        query = collection.query
+        query[1]['v'].append(REAL_ORG_UID_PATTERN.format(self.vendors_uid))
+        collection.setQuery(query)
+        self.assertEqual(auto_item.adapted().getFinanceAdviceId(), self.vendors_uid)
+        self.assertTrue(auto_item.adapted().showFinanceAdviceTemplate())
 
         # if the collection does not exist, [] is returned
         self.deleteAsManager(collection.UID())